optimize keccak256 with >50% less layers (#69)
* optimize keccak256 with ~50% less layers

* fix potiential inbalance layer assign which cause quite lots of paste_from constraints
