+++ gkr/Makefile.toml
@@ -5,13 +5,13 @@ CARGO_MAKE_EXTEND_WORKSPACE_MAKEFILE = true
 
 [tasks.gkr_bench]
 command = "cargo"
-args = ["bench", "--bench", "keccak256", "--features", "flamegraph", "--package", "gkr"]
+args = ["bench", "--bench", "keccak256", "--features", "flamegraph", "--features", "parallel", "--package", "gkr"]
 
 [tasks.gkr_example]
 command = "cargo"
-args = ["run", "--package", "gkr", "--release", "--example", "keccak256"]
+args = ["run", "--package", "gkr", "--release", "--features", "parallel", "--example", "keccak256"]
 
 [tasks.gkr_example_flamegraph]
 env = { "RUST_LOG" = "debug"}
 command = "cargo"
-args = ["run", "--package", "gkr", "--release", "--example", "keccak256"]
+args = ["run", "--package", "gkr", "--release", "--features", "parallel", "--example", "keccak256"]
