+++ gkr/src/lib.rs
@@ -2,6 +2,7 @@
 
 mod circuit;
 pub mod error;
+pub mod gadgets;
 pub mod macros;
 mod prover;
 pub mod structs;
